tudo atualizado seguindo um padrão
